Create track_transitions.lp
The logic of the track transitions in the original (notebook) file have been tested and are shown to be correct.

This file is a compilation of all of those in one place.  The benefit is that redundancies have been eliminated (e.g. Types 1–5 all have the same type of straight track, so one single line applies for all types).

Two things must still be done:
1. Include Track Type #7 (dead end)
2. Apply line 4 to a non-continuous range (2, 4, 6, 12)