Use raw string literals when specifying regexp.
This fixes the syntax error on current python 3 version. The error is
due to that fact that string literals contain unicode characters. In
unicode, there is a fixed set of valid escape sequences - and \d is
not one of them, so \d is interpreted as in invalid unicode escape
seqence.

The best solution is to use a raw string literal r"\d" - here
backslashes are interpreted as just that, instead of as the beginning
of a unicode escape sequence - and thus we can safely use raw string
literals to write regular expressions.